Improve handling of status messages.
Write all status messages to standard error instead of standard output.
Standard output is reserved for actual output from the program and is
not to be used for errors or progress updates.

Add ``--quiet`` and ``--verbosity`` options to control what messages get
printed. Currently, this is all or nothing, but can be expanded in the
future.

Refactor the ``Main`` class to make it easier to print messages based on
the verbosity level.

Resolves issue #3.