Set version to 0.1.0.
The initial version in the build file, 1.0.0, was set more on a whim
than for any other logical reason. As the program will now be using a
more stringent release cycle, it's best to set the version number
appropriately.

The upcoming release will be the first release, but the program is still
in an "alpha" stage so 0.1.0 seems like a good number.