Move option parsing to ``Config`` class.
The option parsing code is a better fit for the ``Config`` class than
the ``Main`` class. Additionally, this allows some of the details of the
parser implementation to be hidden inside the ``Config`` class (such as
the use of the builder pattern in constructing the ``Config`` object).
Also, this allows values to be initialized to an "not set" state to make
it easier to use more complex default values.

Resolves issue #1.